Update README.md
To launch the application run inside of the Docker container via noVNC (that is, running in your browser). 
In order to run, navigate via the terminal to the folder containing the makefile and run $ make mgba
Use arrow keys to run the Goodman in different directions. Once you successfully win the game, enter backspace to restart.